K-Means and Auto-encoder on Cifar 10 Dataset
The implementation has two parts, first part is K-Means Clustering Implementation from Scratch on the Cifar-10 Dataset and other part is using Auto-encoder to remove the unnecessary feature's of data with Encoding Layer and use this Data to clustering with K-Means Algorithm. Both the Implementation uses the Average Silhouette and Dunn's Index to check the Quality of Clusters Formed.